legacy: WebUser removal
* INCOMPATIBLE Removes legacy WebUser module. (addresses #3233)
